requirements.txt
numpy
scipy
matplotlib
pillow
opencv
